fix: move mixin to separate file
